+++ bboss-elasticsearch-rest/src/main/java/org/frameworkset/elasticsearch/ElasticSearchHelper.java
@@ -162,6 +162,12 @@ public class ElasticSearchHelper {
 				elasticsearchPropes.put("elasticsearch.slowDslCallback",
 						slowDslCallback);
 			}
+			String useHttps = ElasticSearchHelper._getStringValue(serverName,"elasticsearch.useHttps",configContext,null);
+			if(useHttps != null) {
+				elasticsearchPropes.put("elasticsearch.useHttps",
+						useHttps);
+			}
+