+++ bboss-elasticsearch-rest/src/main/java/org/frameworkset/elasticsearch/template/TemplateMeta.java
@@ -1,0 +1,32 @@
+package org.frameworkset.elasticsearch.template;
+/**
+ * Copyright 2008 biaoping.yin
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * <p>Description: </p>
+ * <p></p>
+ * <p>Copyright (c) 2018</p>
+ * @Date 2020/1/17 14:42
+ * @author biaoping.yin
+ * @version 1.0
+ */
+public interface TemplateMeta {
+	Object getExtendAttribute(String extendAttribute);
+
+	Object getObject();
+
+	boolean getBooleanExtendAttribute(String extendAttribute, boolean defaultValue);
+}